requirements.txt
flask
opencv-python
numpy
scipy
pillow
